creacion de helpers para formateos de variables en general
